Add triton tip and move BASE_URL in examples
In certain cases, invocation of vLLM can result in mysterious errors triggered by bad state in the triton cache. Scrubbing the cache resolves this, and the documentation has been updated to reflect that.

Additionally, BASE_URL sitting in the middle of the examples make it prone to be truncated if the serving and inference is split up; in particular due to the inference bits being the only part that references BASE_URL. Passing this as nil does not make litellm fail hard, and instead makes it so that it will use a fallback base (http://localhost:8000) instead of http://localhost:8000/v1.